requirements.txt
# production
aiohttp
aiohttp[speedups]
protobuf
pyopenssl


# test / dev
pylint
pytest
pytest-aiohttp
black
trustme